LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-overflow-002-expected.txt














PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="60"><span></span><br><span></span></div>
offsetLeft expected 60 but got 100
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="100"><span></span><br><span></span></div>
offsetLeft expected 100 but got 270
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="120"><span></span><br><span></span></div>
offsetLeft expected 120 but got 100
PASS .target > * 8
PASS .target > * 9
FAIL .target > * 10 assert_equals:
<div class="inner" data-offset-x="40">
    <div style="margin-block-start: 200px;">
      <span></span><br><span></span>
    </div>
  </div>
offsetLeft expected 40 but got 160
FAIL .target > * 11 assert_equals:
<div data-offset-x="10"><span></span><br><span></span></div>
offsetLeft expected 10 but got 100
PASS .target > * 12
